% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5.tools.R
\name{d2h5_to_array}
\alias{d2h5_to_array}
\title{Transform HDF5 numeric matrix to array}
\usage{
d2h5_to_array(h5handle, dpath)
}
\arguments{
\item{h5handle}{HDF5 file handle, Get handle using h5::h5file(infile, 'r')}

\item{dpath}{character, Path to the HDF5 matrix like data element}
}
\value{
Data in 'dpath' as a numeric array
}
\description{
Transform HDF5 numeric matrix to array
}
