% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools.R
\name{file_body}
\alias{file_body}
\title{Extract filename without extension.}
\usage{
file_body(filename)
}
\arguments{
\item{filename}{character, Full file name}
}
\value{
File name body without extension
}
\description{
Extract filename without extension. For '/a/b.c' returns 'b'.
}
\examples{
file_body('/a/b.c')

}
