% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools.R
\name{file_ext}
\alias{file_ext}
\title{Get file extension from filename}
\usage{
file_ext(filename)
}
\arguments{
\item{filename}{character, Full file name}
}
\value{
File extension from filename
}
\description{
Get file extension from filename. For '/a/b.c' returns 'c'.
}
\examples{
file_ext('/a/b.c')

}
