% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5.tools.R
\name{load.h5.array}
\alias{load.h5.array}
\title{Load data from HDF5 file}
\usage{
load.h5.array(infile, element)
}
\arguments{
\item{infile}{character, Full path to the h5 file}

\item{element}{character, h5 data element to load}
}
\value{
Element 'element' from file as a numeric array
}
\description{
Load data from HDF5 file
}
