% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5.tools.R
\name{load.h5.erp}
\alias{load.h5.erp}
\title{Load all single-trial ERPs from HDF5 file}
\usage{
load.h5.erp(infile)
}
\arguments{
\item{infile}{character, Full file name of the ERP HDF5 file}
}
\value{
A list of single-trial ERP data matrices
}
\description{
Load all single-trial ERPs from HDF5 file
Note: assumes structure /erp/<dataset id> use loaddir.h5*() to
load from single files.
}
