% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5.tools.R
\name{load.h5.erpavg}
\alias{load.h5.erpavg}
\title{Load all average ERPs from HDF5 file}
\usage{
load.h5.erpavg(infile)
}
\arguments{
\item{infile}{character, Full file name of the ERP HDF5 file}
}
\value{
A list of erpavg data matrices
}
\description{
Load all average ERPs from HDF5 file
Note: assumes structure /erpavg/<dataset id> use loaddir.h5*() to
load from single files.
}
