% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_utils.R
\name{load_ctap_mc}
\alias{load_ctap_mc}
\title{Load CTAP MC file contents}
\usage{
load_ctap_mc(mc_file, measurement_sheet = "mc", subject_sheet = "subject",
  blocks_sheet = "blocks")
}
\arguments{
\item{mc_file}{character, Full path to the MC file}

\item{measurement_sheet}{character, Name of sheet that lists measurements}

\item{subject_sheet}{character, Name of sheet that lists subjects}

\item{blocks_sheet}{character, Name of sheet that lists blocks}
}
\value{
A list with elements
        file: mc_file
        subject: subject_sheet contents as tibble
        measurement: measurement_sheet contents as tibble
        blocks: blocks_sheet contents as tibbble
}
\description{
Load CTAP MC file contents
}
