% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_utils.R
\name{load_ctap_mc_blocks}
\alias{load_ctap_mc_blocks}
\title{Load CTAP MC blocks sheet/table}
\usage{
load_ctap_mc_blocks(mc_file, blocks_sheet, timeformat = "\%Y\%m\%dT\%H\%M\%S",
  tz = "UTC")
}
\arguments{
\item{mc_file}{character, Full path to the MC file}

\item{blocks_sheet}{character, Name of sheet that lists blocks}

\item{timeformat}{character, Format string for time fields}

\item{tz}{character, Time zone to use}
}
\value{
A tibble with the contents of the blocks_sheet sheet/table
}
\description{
Load CTAP MC blocks sheet/table. File can be xlsx or sqlite.
}
