% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_utils.R
\name{load_ctap_mc_subject}
\alias{load_ctap_mc_subject}
\title{Load CTAP MC subject sheet/table}
\usage{
load_ctap_mc_subject(mc_file, subject_sheet)
}
\arguments{
\item{mc_file}{character, Full path to the MC file}

\item{subject_sheet}{character, Name of sheet that lists subjects}
}
\value{
A tibble with the contents of the subject_sheet sheet/table
}
\description{
Load CTAP MC subject sheet/table. File can be xlsx or sqlite.
}
