% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5.tools.R
\name{loadfl.h5}
\alias{loadfl.h5}
\title{Load data from an array of hdf5files, output a list of numeric arrays}
\usage{
loadfl.h5(h5_file_list, element, element_names = NULL)
}
\arguments{
\item{h5_file_list}{[1,M] character, An array of full file names}

\item{element}{character, Data element to load}

\item{element_names}{[1,M] character, Optional names for the list elements}
}
\value{
A named list of data matrices
}
\description{
Load data from an array of hdf5files, output a list of numeric arrays
}
\examples{
\dontrun{
h5_file_list <- c('file1.h5','file2.h5')
element <- '/erp'
loadfl.h5(h5_file_list, element)
}

}
