% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools.R
\name{make.ggplot.df}
\alias{make.ggplot.df}
\title{Transform a matrix into a ggplottable data.frame}
\usage{
make.ggplot.df(dmat, dataset = NULL, variable2numeric = T)
}
\arguments{
\item{dmat}{[n,m] numeric, input data matrix}

\item{dataset}{character, data set id string}

\item{variable2numeric}{logical, Should $variable be converted to numeric?}
}
\value{
A data.frame representation of dmat
}
\description{
Converts a numeric matrix into a ggplot compatible data.frame.
}
