% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools.R
\name{parse.savename}
\alias{parse.savename}
\title{CTAP/BWRC savename parsing}
\usage{
parse.savename(fname)
}
\arguments{
\item{fname}{character, Full file name}
}
\value{
A list of elements parsed from file name
}
\description{
Parse a CTAP/BWRC casename, often used in filenames when saving data.
}
