% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools.R
\name{shifter}
\alias{shifter}
\title{Circularly shift vector elements}
\usage{
shifter(x, n)
}
\arguments{
\item{x}{[1,m] numeric, input data}

\item{n}{[1,1] ind, number of positions to shift,
positive n shifts things to the left, negative to the right}
}
\value{
A shifted version of x
}
\description{
To e.g. change the position of a peak within a series
positive n shifts things to the left, negative to the right

From: http://stackoverflow.com/questions/30542128/circular-shifting-arrays-in-r-by-distance-n
}
